+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/TestCompiler.kt
@@ -1,13 +1,21 @@
 package org.jetbrains.research.testspark.core.test
 
 import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
+import org.jetbrains.research.testspark.core.test.strategies.TestCompilerStrategy
+import org.jetbrains.research.testspark.core.utils.DataFilesUtil
+import org.jetbrains.research.testspark.core.utils.Language
 
 data class TestCasesCompilationResult(
     val allTestCasesCompilable: Boolean,
     val compilableTestCases: MutableSet<TestCaseGeneratedByLLM>,
 )
 
-interface TestCompiler {
+class TestCompiler(
+    private val javaHomeDirectoryPath: String,
+    private val libPaths: List<String>,
+    private val junitLibPaths: List<String>,
+    private val language: Language,
+) {
     fun compileTestCases(
         generatedTestCasesPaths: List<String>,
         buildPath: String,
@@ -35,7 +43,20 @@ interface TestCompiler {
      * @return A pair containing a boolean value indicating whether the compilation was successful (true) or not (false),
      *         and a string message describing any error encountered during compilation.
      */
-    fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String>
+    fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String> {
+        return when (language) {
+            Language.Java -> TestCompilerStrategy.compileJavaCode(
+                path,
+                "\"${getClassPaths(projectBuildPath)}\"",
+                javaHomeDirectoryPath
+            )
+
+            Language.Kotlin -> TestCompilerStrategy.compileKotlinCode(
+                path,
+                "\"${getClassPaths(projectBuildPath)}\""
+            )
+        }
+    }
 
     /**
      * Generates the path for the command by concatenating the necessary paths.
@@ -43,5 +64,15 @@ interface TestCompiler {
      * @param buildPath The path of the build file.
      * @return The generated path as a string.
      */
-    fun getPath(buildPath: String): String
+    fun getClassPaths(buildPath: String): String {
+        // create the path for the command
+        val separator = DataFilesUtil.classpathSeparator
+        val dependencyLibPath = libPaths.joinToString(separator.toString())
+        val junitPath = junitLibPaths.joinToString(separator.toString())
+
+        val path = "$junitPath${separator}$dependencyLibPath${separator}$buildPath"
+        println("[TestCompiler]: the path is: $path")
+
+        return path
+    }
 }
