--- core/src/main/kotlin/org/jetbrains/research/testspark/core/test/java/JavaJUnitTestSuiteParser.kt
@@ -1,24 +1,0 @@
-package org.jetbrains.research.testspark.core.test.java
-
-import org.jetbrains.research.testspark.core.data.JUnitVersion
-import org.jetbrains.research.testspark.core.test.TestSuiteParser
-import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
-import org.jetbrains.research.testspark.core.test.strategies.JUnitTestSuiteParserStrategy
-import org.jetbrains.research.testspark.core.utils.Language
-
-class JavaJUnitTestSuiteParser(
-    private val packageName: String,
-    private val junitVersion: JUnitVersion,
-    private val importPattern: Regex,
-) : TestSuiteParser {
-    override fun parseTestSuite(rawText: String): TestSuiteGeneratedByLLM? {
-        return JUnitTestSuiteParserStrategy.parseTestSuite(
-            rawText,
-            junitVersion,
-            importPattern,
-            packageName,
-            testNamePattern = "void",
-            Language.Java,
-        )
-    }
-}
