+++ src/main/kotlin/org/jetbrains/research/testspark/tools/TestCompilerFactory.kt
@@ -4,8 +4,6 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
 import org.jetbrains.research.testspark.core.data.JUnitVersion
 import org.jetbrains.research.testspark.core.test.TestCompiler
-import org.jetbrains.research.testspark.core.test.java.JavaTestCompiler
-import org.jetbrains.research.testspark.core.test.kotlin.KotlinTestCompiler
 import org.jetbrains.research.testspark.core.utils.Language
 
 class TestCompilerFactory {
@@ -23,10 +21,7 @@ class TestCompilerFactory {
             val libraryPaths = LibraryPathsProvider.getTestCompilationLibraryPaths()
             val junitLibraryPaths = LibraryPathsProvider.getJUnitLibraryPaths(junitVersion)
 
-            return when (language) {
-                Language.Java -> JavaTestCompiler(javaSDKHomePath, libraryPaths, junitLibraryPaths)
-                Language.Kotlin -> KotlinTestCompiler(libraryPaths, junitLibraryPaths)
-            }
+            return TestCompiler(javaSDKHomePath, libraryPaths, junitLibraryPaths, language)
         }
     }
 }
